+++ src/rx-schema.js
@@ -182,7 +182,15 @@ export class RxSchema {
      * returns true if key-compression should be done
      */
     doKeyCompression() {
-        return !this.jsonID.disableKeyCompression;
+        /**
+         * in rxdb 8.0.0 we renambed the keycompression-option
+         * But when a data-migration is done with and old schema,
+         * it might have the old option which then should be used
+         * TODO: Remove this check in Sep 2019
+         */
+        if (this.jsonID.hasOwnProperty('disableKeyCompression')) {
+            return !this.jsonID.disableKeyCompression;
+        } else return this.jsonID.keyCompression;
     }
 
     /**
@@ -316,8 +324,8 @@ const fillWithDefaultSettings = function(schemaObj) {
     schemaObj.additionalProperties = false;
 
     // fill with key-compression-state ()
-    if (!schemaObj.hasOwnProperty('disableKeyCompression'))
-        schemaObj.disableKeyCompression = true;
+    if (!schemaObj.hasOwnProperty('keyCompression'))
+        schemaObj.keyCompression = false;
 
     // compoundIndexes must be array
     schemaObj.compoundIndexes = schemaObj.compoundIndexes || [];