+++ test/unit/core.node.js
@@ -16,7 +16,7 @@ const schema = {
     title: 'human schema',
     description: 'describes a human being',
     version: 0,
-    disableKeyCompression: true,
+    keyCompression: false,
     type: 'object',
     properties: {
         passportId: {